Backport total_memory improvements from anaconda
The version used is a merger of the original blivet code and anaconda's
counterpart from commit f22386ef. Note that the referenced commit has not been
merged, and so isn't reachable in the repository itself. See below for
relevant GitHub links. Note also that the function in anaconda has been
retired in favor of using the blivet's one.

New:
- checks for bad input
- check for no data
- tests

See also:
- https://github.com/rhinstaller/anaconda/pull/4338
- https://github.com/rhinstaller/anaconda/commit/f22386efafe33df7f8182e65aaf99279904344c6